fix NumberFormat problem with SportsApp
The background seems to be that US uses "." as a decimal seperator and Europe ","
When the language is changed to a European language (or maybe also others) then SportsApp crashes because it no longer understands the data it is storing or has stored
The proposed solution is to keep the language at "en", but change the region. This apparently does not affect the NumberFormat used.